CADC-10534 - unit tests are working again.

+++ cfht2caom2/tests/test_composable.py
@@ -69,8 +69,10 @@
 
 import os
 import shutil
+import warnings
 
 from astropy.io import fits
+from astropy.utils.exceptions import AstropyUserWarning
 from datetime import datetime
 from hashlib import md5
 
@@ -79,6 +81,7 @@ from mock import Mock, patch
 from cadcdata import FileInfo
 from caom2utils import data_util
 from caom2 import SimpleObservation, Algorithm, Instrument
+from caom2pipe import astro_composable as ac
 from caom2pipe import caom_composable as cc
 from caom2pipe import data_source_composable as dsc
 from caom2pipe import manage_composable as mc
@@ -89,28 +92,39 @@ import cfht_mocks
 TEST_DIR = f'{test_main_app.TEST_DATA_DIR}/composable_test'
 
 
+@patch('caom2pipe.astro_composable.check_fits')
+@patch('caom2utils.data_util.get_local_headers_from_fits')
 @patch('cadcutils.net.ws.WsCapabilities.get_access_url')
-@patch('caom2pipe.execute_composable.CaomExecute._fits2caom2_cmd_local')
 @patch('caom2pipe.client_composable.CAOM2RepoClient')
-def test_run_by_builder(repo_mock, exec_mock, access_mock):
-    access_mock.return_value = 'https://localhost'
-    # should attempt to run LocalMetaCreate
-    repo_mock.return_value.read.side_effect = _mock_repo_read
-    repo_mock.return_value.create.side_effect = Mock()
-    repo_mock.return_value.update.side_effect = _mock_repo_update
-    getcwd_orig = os.getcwd
-    os.getcwd = Mock(return_value=TEST_DIR)
-    try:
-        # execution
-        test_result = composable._run_by_builder()
-        assert test_result == 0, 'wrong result'
-    finally:
-        os.getcwd = getcwd_orig
+def test_run_by_builder(
+    repo_mock, access_mock, util_headers_mock, fits_mock
+):
+    util_headers_mock.side_effect = ac.make_headers_from_file
+    # files are valid FITS
+    fits_mock.return_value = True
 
-    assert repo_mock.return_value.read.called, 'repo read not called'
-    assert repo_mock.return_value.create.called, 'repo create not called'
-    assert exec_mock.called, 'expect to be called'
-    exec_mock.assert_called_with(), 'wrong exec args'
+    test_fqn = os.path.join(TEST_DIR, 'test_files')
+    if not os.path.exists(test_fqn):
+        os.mkdir(test_fqn)
+    try:
+        access_mock.return_value = 'https://localhost'
+        # should attempt to run LocalMetaCreate
+        repo_mock.return_value.read.side_effect = _mock_repo_read
+        repo_mock.return_value.create.side_effect = Mock()
+        repo_mock.return_value.update.side_effect = _mock_repo_update
+        getcwd_orig = os.getcwd
+        os.getcwd = Mock(return_value=TEST_DIR)
+        try:
+            # execution
+            test_result = composable._run_by_builder()
+            assert test_result == 0, 'wrong result'
+        finally:
+            os.getcwd = getcwd_orig
+
+        assert repo_mock.return_value.read.called, 'repo read not called'
+        assert repo_mock.return_value.create.called, 'repo create not called'
+    finally:
+        _cleanup(TEST_DIR)
 
 
 @patch('caom2pipe.astro_composable.check_fits')
@@ -156,16 +170,18 @@ def test_run_store_retry(
     test_dir_fqn = os.path.join(
         test_main_app.TEST_DATA_DIR, 'store_retry_test'
     )
-    test_failure_fqn = os.path.join(
+    test_failure_dir = os.path.join(
         test_main_app.TEST_DATA_DIR,
-        'store_retry_test/failure/1000003f.fits.fz',
+        'store_retry_test/failure',
     )
-    test_success_fqn = os.path.join(
+    test_success_dir = os.path.join(
         test_main_app.TEST_DATA_DIR, 'store_retry_test/success'
     )
-    for ii in [test_failure_fqn, test_success_fqn]:
+    for ii in [test_failure_dir, test_success_dir]:
         if not os.path.exists(ii):
             os.mkdir(ii)
+    test_failure_fqn = os.path.join(test_failure_dir, '1000003f.fits.fz')
+    test_success_fqn = os.path.join(test_success_dir, '1000003f.fits.fz')
 
     try:
         getcwd_orig = os.getcwd
@@ -195,24 +211,16 @@ def test_run_store_retry(
             f'{test_dir_fqn}/new', 'ad:CFHT/1000003f.fits.fz', 'default'
         ), 'wrong put_file args'
         assert os.path.exists(test_failure_fqn), 'expect failure move'
-        success_content = os.listdir(test_success_fqn)
+        success_content = os.listdir(test_success_dir)
         assert len(success_content) == 0, 'should be no success files'
     finally:
         if os.path.exists(test_failure_fqn):
             test_new_fqn = os.path.join(test_dir_fqn, 'new/1000003f.fits.fz')
             shutil.move(test_failure_fqn, test_new_fqn)
-        for ii in [
-            f'{test_dir_fqn}/logs',
-            f'{test_dir_fqn}/logs_0',
-            f'{test_dir_fqn}/metrics',
-            f'{test_dir_fqn}/rejected',
-        ]:
-            if os.path.exists(ii):
-                for entry in os.scandir(ii):
-                    os.unlink(entry.path)
-                os.rmdir(ii)
+        _cleanup(test_dir_fqn)
 
 
+@patch('caom2utils.data_util.get_local_headers_from_fits')
 @patch('caom2pipe.client_composable.CAOM2RepoClient')
 @patch('caom2pipe.client_composable.CadcTapClient')
 @patch('caom2pipe.client_composable.StorageClientWrapper')
@@ -223,40 +231,48 @@ def test_run_store_retry(
 )
 @patch('caom2pipe.execute_composable.OrganizeExecutes.do_one')
 def test_run_state(
-    run_mock, tap_mock, data_client_mock, tap_client_mock, repo_mock
+    run_mock, tap_mock, data_client_mock, tap_client_mock, repo_mock,
+    util_headers_mock
 ):
-    run_mock.return_value = 0
-    tap_mock.side_effect = _mock_dir_listing
-    data_client_mock.return_value.get_head.side_effect = cfht_mocks._mock_get_head
-    data_client_mock.return_value.info.side_effect = (
-        _mock_get_file_info
-    )
-    getcwd_orig = os.getcwd
-    os.getcwd = Mock(return_value=TEST_DIR)
-
-    test_obs_id = '2359320p'
-    test_f_name = f'{test_obs_id}.fits'
     try:
-        # execution
-        test_result = composable._run_state()
-        assert test_result == 0, 'mocking correct execution'
-    finally:
-        os.getcwd = getcwd_orig
+        util_headers_mock.side_effect = ac.make_headers_from_file
+        run_mock.return_value = 0
+        tap_mock.side_effect = _mock_dir_listing
+        data_client_mock.return_value.get_head.side_effect = (
+            cfht_mocks._mock_get_head
+        )
+        data_client_mock.return_value.info.side_effect = (
+            _mock_get_file_info
+        )
+        getcwd_orig = os.getcwd
+        os.getcwd = Mock(return_value=TEST_DIR)
 
-    assert run_mock.called, 'should have been called'
-    args, kwargs = run_mock.call_args
-    test_storage = args[0]
-    assert isinstance(test_storage, cfht_name.CFHTName), type(test_storage)
-    assert (
-        test_storage.obs_id == test_obs_id
-    ), f'wrong obs id {test_storage.obs_id}'
-    assert test_storage.file_name == test_f_name, 'wrong file name'
-    assert test_storage.fname_on_disk == test_f_name, 'wrong fname on disk'
-    assert test_storage.url is None, 'wrong url'
-    assert (
-        test_storage.lineage == f'{test_obs_id}/ad:CFHT/{test_f_name}'
-    ), 'wrong lineage'
-    assert test_storage.external_urls is None, 'wrong external urls'
+        # it's a WIRCAM file
+        test_obs_id = '2281792'
+        test_f_name = f'{test_obs_id}p.fits.fz'
+        try:
+            # execution
+            test_result = composable._run_state()
+            assert test_result == 0, 'mocking correct execution'
+        finally:
+            os.getcwd = getcwd_orig
+
+        assert run_mock.called, 'should have been called'
+        args, kwargs = run_mock.call_args
+        test_storage = args[0]
+        assert isinstance(test_storage, cfht_name.CFHTName), type(test_storage)
+        assert (
+            test_storage.obs_id == test_obs_id
+        ), f'wrong obs id {test_storage.obs_id}'
+        assert test_storage.file_name == test_f_name, 'wrong file name'
+        assert test_storage.fname_on_disk == test_f_name, 'wrong fname on disk'
+        assert test_storage.url is None, 'wrong url'
+        assert (
+            test_storage.lineage == f'{test_obs_id}p/ad:CFHT/{test_f_name}'
+        ), 'wrong lineage'
+        assert test_storage.external_urls is None, 'wrong external urls'
+    finally:
+        _cleanup(TEST_DIR)
 
 
 @patch('caom2pipe.client_composable.CAOM2RepoClient')
@@ -292,36 +308,63 @@ def test_run_by_builder_hdf5_first(data_mock, tap_mock, repo_mock):
     _common_execution(test_dir, actual_fqn, expected_hdf5_only_fqn)
 
 
+@patch('caom2utils.data_util.get_local_headers_from_fits')
+@patch('caom2pipe.astro_composable.check_fits')
 @patch('caom2pipe.client_composable.CAOM2RepoClient')
 @patch('caom2pipe.client_composable.CadcTapClient')
 @patch('caom2pipe.client_composable.StorageClientWrapper')
-def test_run_by_builder_hdf5_added_to_existing(data_mock, tap_mock, repo_mock):
-    # add to an existing observation with an hdf5 file, just using scrape
-    # to make sure the observation is writable to an ams service, and the
-    # 'p' metadata gets duplicated correctly
+def test_run_by_builder_hdf5_added_to_existing(
+    data_mock, tap_mock, repo_mock, fits_check_mock, header_mock
+):
+    try:
+        warnings.simplefilter('ignore', category=AstropyUserWarning)
+        fits_check_mock.return_value = True
+        header_mock.side_effect = ac.make_headers_from_file
 
-    test_obs_id = '2384125p'
-    test_dir = f'{test_main_app.TEST_DATA_DIR}/hdf5_test'
-    hdf5_fqn = f'{test_dir}/2384125z.hdf5'
-    fits_fqn = f'{test_dir}/{test_obs_id}.fits.header'
-    actual_fqn = f'{test_dir}/logs/{test_obs_id}.xml'
-    expected_fqn = f'{test_dir}/all.expected.xml'
-    expected_hdf5_only_fqn = f'{test_dir}/hdf5_only.expected.xml'
+        # add to an existing observation with an hdf5 file, just using scrape
+        # to make sure the observation is writable to an ams service, and the
+        # 'p' metadata gets duplicated correctly
 
-    # make sure expected files are present
-    shutil.copy(expected_hdf5_only_fqn, actual_fqn)
-    if not os.path.exists(fits_fqn):
-        shutil.copy(
-            f'{test_main_app.TEST_DATA_DIR}/multi_plane/'
-            f'{test_obs_id}.fits.header',
-            fits_fqn,
-        )
+        test_obs_id = '2384125p'
+        test_dir = f'{test_main_app.TEST_DATA_DIR}/hdf5_test'
+        hdf5_fqn = f'{test_dir}/2384125z.hdf5'
+        fits_fqn = f'{test_dir}/{test_obs_id}.fits.header'
+        actual_fqn = f'{test_dir}/logs/{test_obs_id}.xml'
+        expected_fqn = f'{test_dir}/all.expected.xml'
+        expected_hdf5_only_fqn = f'{test_dir}/hdf5_only.expected.xml'
 
-    # clean up unexpected files
-    if os.path.exists(hdf5_fqn):
-        os.unlink(hdf5_fqn)
+        # make sure expected files are present
+        if not os.path.exists(actual_fqn):
+            if not os.path.exists(os.path.join(test_dir, 'logs')):
+                os.mkdir(os.path.join(test_dir, 'logs'))
+            shutil.copy(expected_hdf5_only_fqn, actual_fqn)
+        if not os.path.exists(fits_fqn):
+            shutil.copy(
+                f'{test_main_app.TEST_DATA_DIR}/multi_plane/'
+                f'{test_obs_id}.fits.header',
+                fits_fqn,
+            )
 
-    _common_execution(test_dir, actual_fqn, expected_fqn)
+        # clean up unexpected files
+        if os.path.exists(hdf5_fqn):
+            os.unlink(hdf5_fqn)
+
+        _common_execution(test_dir, actual_fqn, expected_fqn)
+    finally:
+        _cleanup(test_dir)
+
+
+def _cleanup(test_dir_fqn):
+    for ii in [
+        f'{test_dir_fqn}/logs',
+        f'{test_dir_fqn}/logs_0',
+        f'{test_dir_fqn}/metrics',
+        f'{test_dir_fqn}/rejected',
+    ]:
+        if os.path.exists(ii):
+            for entry in os.scandir(ii):
+                os.unlink(entry.path)
+            os.rmdir(ii)
 
 
 def _common_execution(test_dir, actual_fqn, expected_fqn):
@@ -397,7 +440,7 @@ def _mock_dir_listing(
     return [
         dsc.StateRunnerMeta(
             os.path.join(
-                os.path.join(TEST_DIR, 'test_files'), '2359320p.fits'
+                os.path.join(TEST_DIR, 'test_files'), '2281792p.fits.fz'
             ),
             '2019-10-23T16:27:19.000',
         ),
